+++ packages/form/src/TextField/TextField.js
@@ -29,8 +29,10 @@ type Props = {
   mask?: string,
   state?: InputStates,
   autoComplete?: string,
-  iconright?: string,
-  iconleft?: string,
+  iconright?: React.ReactElement | string,
+  iconleft?: React.ReactElement | string,
+  addonleft?: string,
+  addonright?: string,
   loading?: boolean,
   type?: InputTypes;
   /** Qa id */
@@ -56,6 +58,8 @@ class TextField extends Component<Props> {
       qa,
       type,
       errorDescription,
+      addonleft,
+      addonright,
       ...extraProps,
     } = this.props;
 
@@ -64,6 +68,7 @@ class TextField extends Component<Props> {
       {
         'has-icon-right': iconright || loading,
         'has-icon-left': iconleft,
+        'has-addon': !!addonleft || !!addonright,
         'is-required': !!required,
         [`${stateClasses[state]}`]: !!state,
         [`a-input--${type}`]: !!type,
@@ -74,12 +79,14 @@ class TextField extends Component<Props> {
       <div className={inputClass} data-qa={qa}>
         {label && <InputLabel htmlFor={id}>{label}</InputLabel>}
         <div className="a-input__wrapper">
+          {addonleft ? <div class="a-input__addon">{addonleft}</div> : null}
           {iconleft ? <Icon name={iconleft}/> : null}
           {mask ? <InputMask {...this.props} /> :
             <input id={id} type="text" autoComplete={autoComplete} disabled={disabled}
                    ref={ref => this.props.inputRef && this.props.inputRef(ref)} {...extraProps} />}
           {iconright ? <Icon name={iconright}/> : null}
           {loading ? <span className="fa a-spinner a-spinner--sm"/> : null}
+          {addonright ? <div class="a-input__addon">{addonright}</div> : null}
         </div>
         {description ? <small>{description}</small> : null}
         {errorDescription && <small className="u-text-danger">{errorDescription}</small>}