+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+- `Form/TextField` Support addons (addon-left, addon-right)
 
 ### Fixed
 - `Autocomplete` Fixed a bug where search results were not correctly parsed